requirements.txt
openai
python-dotenv
IPython
opencv-python
requests
pytube
yt-dlp
